:art: refactor: move `DuckDbPrecalcQueryService`
