+++ api/app/routers/land_change/tree_cover_gain/tree_cover_gain.py
@@ -3,7 +3,6 @@ from app.domain.analyzers.dummy_tree_cover_gain_analyzer import (
 )
 from app.domain.compute_engines.compute_engine import (
     ComputeEngine,
-    DuckDbPrecalcQueryService,
     FloxOTFHandler,
     PrecalcQueryBuilder,
     TreeCoverGainPrecalcHandler,
@@ -13,6 +12,9 @@ from app.domain.repositories.data_api_aoi_geometry_repository import (
     DataApiAoiGeometryRepository,
 )
 from app.domain.repositories.zarr_dataset_repository import ZarrDatasetRepository
+from app.infrastructure.external_services.duck_db_query_service import (
+    DuckDbPrecalcQueryService,
+)
 from app.infrastructure.persistence.file_system_analysis_repository import (
     FileSystemAnalysisRepository,
 )
